--- src/main/resources/socket.md
@@ -1,19 +1,0 @@
-# Socket
-
-Создание файлового сервера и **CLI** (*Command Line Interface*) для клиента файлового сервера.
-
-## Task 1
-
-Добавить следующую команд для файлового сервера:
-- `LIST` - вернуть список доступных файлов
-- `GET ${filepath}` - вернуть файл по указанному `filepath` клиенту или `FILE NOT EXIST`, если такого файла не существует.
-- `SEND ${filepath}` - создать файл клиента по указанному `filepath` на сервере, в случае успеха вернуть - `OK`, в случае неудачи - `FILE EXIST`
-- `DELETE ${filepath}` - удалить файл по указанному `filepath` на сервере,  в случае успеха вернуть - `OK`, в случае неудачи - `FILE NOT EXIST`
-
-## Task 2
-
-Сервер должен логировать все операции в консоль.
-
-## Task 3
-
-Создать **CLI** для клиента, где он мог бы указать `host` и `port` файлового хранилища. После чего, выбрать команду и, если необходимо, указать `filepath`.
